+++ qbraid/_logging.py
@@ -13,7 +13,7 @@ Configures the logging for the qBraid-SDK.
 
 The log level is set based on the `LOG_LEVEL` environment variable, which can be a string
 (e.g., "DEBUG", "INFO") or an integer (e.g., 10 for DEBUG). If an invalid value is provided,
-the log level defaults to `INFO`, and a warning is logged.
+the log level defaults to `WARNING`, and a warning is logged.
 
 Format: `%(levelname)s - %(message)s`
 
@@ -32,7 +32,7 @@ DEFAULT_LOG_LEVEL = logging.WARNING
 
 
 def get_log_level_from_env():
-    """Get the log level from the LOG_LEVEL environment variable, or default to INFO."""
+    """Get the log level from the LOG_LEVEL environment variable, or default to WARNING."""
     log_level_env = os.getenv("LOG_LEVEL", "")
 
     if log_level_env.isdigit():
@@ -41,7 +41,7 @@ def get_log_level_from_env():
             return log_level
 
         logging.warning(
-            "Invalid log level (int) in LOG_LEVEL: %s. Falling back to INFO.", log_level
+            "Invalid log level (int) in LOG_LEVEL: %s. Falling back to WARNING.", log_level
         )
         return DEFAULT_LOG_LEVEL
 
@@ -53,7 +53,7 @@ def get_log_level_from_env():
 
     if log_level_env:
         logging.warning(
-            "Invalid log level (str) in LOG_LEVEL: %s. Falling back to INFO.", log_level_env
+            "Invalid log level (str) in LOG_LEVEL: %s. Falling back to WARNING.", log_level_env
         )
 
     return DEFAULT_LOG_LEVEL