+++ qbraid/runtime/native/job.py
@@ -18,22 +18,18 @@ from typing import TYPE_CHECKING, Optional
 
 from qbraid_core.services.quantum import QuantumClient
 
+from qbraid._logging import logger
 from qbraid.runtime.enums import JobStatus
-from qbraid.runtime.exceptions import JobStateError
+from qbraid.runtime.exceptions import JobStateError, QbraidRuntimeError
 from qbraid.runtime.job import QuantumJob
 from qbraid.runtime.result import GateModelResultData, Result
-
-from .result_data import (
-    NECVectorAnnealerResultData,
-    QbraidQirSimulatorResultData,
-    QuEraQasmSimulatorResultData,
-)
-from .schemas import (
-    NECVectorAnnealerMetadata,
+from qbraid.runtime.schemas import (
     QbraidQirSimulationMetadata,
     QuEraQasmSimulationMetadata,
     RuntimeJobModel,
 )
+
+from .result import  NECVectorAnnealerResultData, QbraidQirSimulatorResultData, QuEraQasmSimulatorResultData
 
 if TYPE_CHECKING:
     import qbraid_core.services.quantum
@@ -68,29 +64,18 @@ class QbraidJob(QuantumJob):
             self._client = self._device.client if self._device else QuantumClient()
         return self._client
 
-    @staticmethod
-    def _map_status(status: Optional[str]) -> JobStatus:
-        """Returns `JobStatus` object mapped from raw status value. If no value
-        provided or conversion from string fails, returns `JobStatus.UNKNOWN`."""
-        if status is None:
-            return JobStatus.UNKNOWN
-        if not isinstance(status, str):
-            raise ValueError(f"Invalid status value type: {type(status)}")
-        for e in JobStatus:
-            status_enum = JobStatus(e.value)
-            if status == status_enum.name or status == str(status_enum):
-                return status_enum
-        raise ValueError(f"Status value '{status}' not recognized.")
-
     def status(self) -> JobStatus:
         """Return the status of the job / task , among the values of ``JobStatus``."""
-        job_data = self.client.get_job(self.id)
-        status: Optional[str] = job_data.get("status")
-        status_msg: Optional[str] = job_data.get("statusText")
-        status_enum = self._map_status(status)
-        if status_msg is not None:
-            status_enum.set_status_message(status_msg)
-        return status_enum
+        terminal_states = JobStatus.terminal_states()
+        if self._cache_metadata.get("status") not in terminal_states:
+            client_data = self.client.get_job(self.id)
+            job_model = RuntimeJobModel.from_dict(client_data)
+            job_data = job_model.model_dump(exclude={"metadata", "cost"})
+            status = JobStatus(job_data.pop("status"))
+            if job_model.status_text is not None:
+                status.set_status_message(job_model.status_text)
+            self._cache_metadata.update({**job_data, "status": status})
+        return self._cache_metadata["status"]
 
     def cancel(self) -> None:
         """Attempt to cancel the job."""
@@ -98,6 +83,19 @@ class QbraidJob(QuantumJob):
             raise JobStateError("Cannot cancel job in a terminal state.")
 
         self.client.cancel_job(self.id)
+        logger.info("Cancel job request validated.")
+
+        try:
+            logger.info("Waiting for job to cancel...")
+            self.wait_for_final_state(timeout=3, poll_interval=1)
+        except JobStateError:
+            pass
+
+        status = self.status()
+        if status not in {JobStatus.CANCELLED, JobStatus.CANCELLING}:
+            raise QbraidRuntimeError(f"Failed to cancel job. Current status: {status.name}")
+
+        logger.info("Success. Current status: %s", status.name)
 
     def result(self) -> Result:
         """Return the results of the job."""